backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
httpx==0.27.0
openai==1.23.3
yfinance==0.2.40
pymongo[srv]==4.6.3